backend/requirements.txt
fastapi
uvicorn
pandas
numpy

scikit-learn
joblib
scipy
firebase-admin
lightgbm
torch
groq
python-dotenv
tensorflow-cpu